+++ lib/src/main/java/com/bartnik/eventstore/communication/EventListenerFactory.java
@@ -1,0 +1,9 @@
+package com.bartnik.eventstore.communication;
+
+import com.bartnik.eventstore.execution.registry.aggregate.EventHandlerDescriptor;
+
+public interface EventListenerFactory {
+    
+    EventListener forEvent(EventHandlerDescriptor descriptor);
+    
+}
