--- lib/src/main/java/com/bartnik/eventstore/execution/communication/EventListenerFactory.java
@@ -1,9 +1,0 @@
-package com.bartnik.eventstore.execution.communication;
-
-import com.bartnik.eventstore.registry.aggregate.EventHandlerDescriptor;
-
-public interface EventListenerFactory {
-    
-    EventListener forEvent(EventHandlerDescriptor descriptor);
-    
-}
